+++ src/app/material/material-index/material-index.component.ts
@@ -68,7 +68,7 @@ export class MaterialIndexComponent implements OnInit, OnDestroy {
           dialogConfig.disableClose = true;
           dialogConfig.autoFocus = true;
           dialogConfig.width = '1000px';
-          dialogConfig.height = '80vh';
+         // dialogConfig.height = '0vh';
           dialogConfig.position = {'top': '50px'};
           this.dialog.open(MaterialCreationDialogComponent, dialogConfig);
       }