+++ src/app/services/project/project.service.ts
@@ -1,8 +1,7 @@
 import { environment } from '../../../environments/environment';
 import { Injectable } from '@angular/core';
-import { JwtHelperService } from '@auth0/angular-jwt';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { catchError, map, tap, delay, shareReplay, publishReplay, refCount } from 'rxjs/operators';
+import { HttpClient,  } from '@angular/common/http';
+import { catchError, map, tap, } from 'rxjs/operators';
 import { Observable, of, BehaviorSubject, throwError } from 'rxjs';
 import { NotificationsService } from 'angular2-notifications';